// src/proxy/types.ts
export enum ResponseTypes {
    reg = "reg",
    create_room = "create_room",
    add_ship = "add_ships",
    update_room = "update_room",
    add_user_to_room = "add_user_to_room",
    attack = 'attack',
    start_game = "start_game",
    turn = "turn",
    randomAttack = "randomAttack",
    finish = "finish",
    single_play = "single_play",
}